# my-app/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Employee & Attendance API
  description: API for managing employees and their attendance records.
  version: 1.0.0

servers:
  - url: http://localhost:3000

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Employee:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Attendance:
      type: object
      properties:
        id:
          type: integer
        employee_id:
          type: integer
        date:
          type: string
          format: date
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

security:
  - bearerAuth: []

paths:
  /api/employee:
    post:
      summary: Create new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        201:
          description: Employee created
        500:
          description: Server error

    get:
      summary: Get all employees
      responses:
        200:
          description: List of employees

  /api/employee/{id}:
    get:
      summary: Get employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Employee found
        404:
          description: Not found

    put:
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        200:
          description: Employee updated

    delete:
      summary: Delete employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Employee deleted

  /api/attendance/auth:
    post:
      summary: Create attendance (JWT required)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Attendance'
      responses:
        201:
          description: Attendance created

  /api /attendance:
    get:
      summary: Get all attendance (JWT required)
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of attendance records

  /api/attendance/limit:
    get:
      summary: Get attendance with pagination (JWT required)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        200:
          description: Paginated attendance list

  /api/attendance/employee/{employeeId}:
    get:
      summary: Get attendance by employee ID (JWT required)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: employeeId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Attendance records

  /api/attendance/{id}:
    put:
      summary: Update attendance (JWT required)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
      responses:
        200:
          description: Attendance updated

    delete:
      summary: Delete attendance (JWT required)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Attendance deleted